# VitruviusServer-openapi.yaml
openapi: "3.1.0"
info:
  title: "VitruviusServer API"
  description: "VitruviusServer API"
  version: "1.0.0"
servers:
  - url: "https://VitruviusServer"
paths:
  /api/v1/vsums:
    get:
      summary: "GET api/v1/vsums"
      description: "Retrieves an array of VSUM information records, containing details like ID, metamodel name, name, and description."
      operationId: "getVsums"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/VsumInfoResponseBody"
    post:
      summary: "POST api/v1/vsums"
      description: "Creates a new VSUM information record based on the provided metamodel name, name, and description."
      operationId: "createVsum"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVsumRequestBody"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/VsumInfoResponseBody"
  /api/v1/vsums/{vsumId}:
    put:
      summary: "PUT api/v1/vsums/{vsumId}"
      description: "Updates an existing VSUM information record (name and description) identified by the provided VSUM ID."
      operationId: "updateVsumInfo"
      parameters:
        - name: "vsumId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVsumInfoRequestBody"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/VsumInfoResponseBody"
    delete:
      summary: "DELETE api/v1/vsums/{vsumId}"
      description: "Deletes the VSUM information record identified by the provided VSUM ID."
      operationId: "deleteVsum"
      parameters:
        - name: "vsumId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      responses:
        "200":
          description: "OK"
  /api/v1/vsums/{vsumId}/view-types:
    get:
      summary: "GET api/v1/vsums/{vsumId}/view-types"
      description: "Retrieves the available view types associated with the specified VSUM ID."
      operationId: "getViewTypes"
      parameters:
        - name: "vsumId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  type: "string"
  /api/v1/vsums/{vsumId}/view-types/{viewTypeName}/selectors:
    post:
      summary: "POST api/v1/vsums/{vsumId}/view-types/{viewTypeName}/selectors"
      description: "Creates a selector for a specified VSUM and view type, returning a JSON string containing the selector identifier (id) and a list of selectableObjects."
      operationId: "createSelector"
      parameters:
        - name: "vsumId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
        - name: "viewTypeName"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "string"
...
components:
  schemas:
    VsumInfoResponseBody:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
          nullable: true
        metamodelName:
          type: "string"
          nullable: true
        name:
          type: "string"
          nullable: true
        description:
          type: "string"
          nullable: true
    CreateVsumRequestBody:
      type: "object"
      properties:
        metamodelName:
          type: "string"
          nullable: true
        name:
          type: "string"
          nullable: true
        description:
          type: "string"
          nullable: true
    UpdateVsumInfoRequestBody:
      type: "object"
      properties:
        name:
          type: "string"
          nullable: true
        description:
          type: "string"
          nullable: true
...